requirements.txt
# ============================================================

numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
scikit-learn>=1.3.0

pykeen>=1.10.0

torch-geometric>=2.3.0

pyyaml>=6.0
tqdm>=4.65.0
colorama>=0.4.6

matplotlib>=3.7.0
seaborn>=0.12.0
